Renamed index to app and server to index in the BE. Ammended start script. Deleted dev script. Added dist to gitignore.
