+++ backend/.gitignore
@@ -23,3 +23,6 @@ npm-debug.log*
 
 # Compiled binary addons
 build/Release
+
+# Compiled directory
+dist
