+++ backend/package.json
@@ -6,9 +6,8 @@
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1",
     "build": "npx tsc",
-    "server": "node --watch dist/server.js",
-    "start": "concurrently \"tsc -watch\" \"npm run server\" \"cd ../frontend && npm run dev\"",
-    "dev": "nodemon src/index.ts"
+    "server": "node --watch dist/index.js",
+    "start": "concurrently \"tsc -watch\" \"npm run server\" \"cd ../frontend && npm run dev\""
   },
   "keywords": [],
   "author": "",
